[#2442] improvement(spark-connector): Refactor spark connector package name to keep consistent with trino
