+++ spark-connector/src/main/java/com/datastrato/gravitino/spark/connector/ConnectorConstants.java
@@ -3,7 +3,7 @@
  *  This software is licensed under the Apache License version 2.
  */
 
-package com.datastrato.gravitino.spark;
+package com.datastrato.gravitino.spark.connector;
 
 public class ConnectorConstants {
   public static final String COMMENT = "comment";
