+++ spark-connector/src/main/java/com/datastrato/gravitino/spark/connector/PropertiesConverter.java
@@ -3,7 +3,7 @@
  *  This software is licensed under the Apache License version 2.
  */
 
-package com.datastrato.gravitino.spark;
+package com.datastrato.gravitino.spark.connector;
 
 import java.util.Map;
