+++ spark-connector/src/main/java/com/datastrato/gravitino/spark/connector/hive/HiveAdaptor.java
@@ -3,13 +3,13 @@
  *  This software is licensed under the Apache License version 2.
  */
 
-package com.datastrato.gravitino.spark.hive;
+package com.datastrato.gravitino.spark.connector.hive;
 
 import com.datastrato.gravitino.rel.Table;
-import com.datastrato.gravitino.spark.GravitinoCatalogAdaptor;
-import com.datastrato.gravitino.spark.GravitinoSparkConfig;
-import com.datastrato.gravitino.spark.PropertiesConverter;
-import com.datastrato.gravitino.spark.table.SparkBaseTable;
+import com.datastrato.gravitino.spark.connector.GravitinoCatalogAdaptor;
+import com.datastrato.gravitino.spark.connector.GravitinoSparkConfig;
+import com.datastrato.gravitino.spark.connector.PropertiesConverter;
+import com.datastrato.gravitino.spark.connector.table.SparkBaseTable;
 import com.google.common.base.Preconditions;
 import java.util.HashMap;
 import java.util.Map;
