+++ spark-connector/src/main/java/com/datastrato/gravitino/spark/connector/plugin/GravitinoSparkPlugin.java
@@ -3,7 +3,7 @@
  *  This software is licensed under the Apache License version 2.
  */
 
-package com.datastrato.gravitino.spark.plugin;
+package com.datastrato.gravitino.spark.connector.plugin;
 
 import org.apache.spark.api.plugin.DriverPlugin;
 import org.apache.spark.api.plugin.ExecutorPlugin;
